At this point, what remains is:
- Adjusting the crate container to be backed by an IItemHandler instead
of an IInventory (and adjust the slots respectively)
- Ensuring that the correct number of people opening is being tracked
for playing sound/displaying the change in visual of the crate
(cross-reference anything barrel-related)
- Updating the various screens to function with the new CrateContainer,
which should theoretically not require much in the way of change.

Overall, some changes, some methods left public made private simply so I
didn't have to scream at intellij.