Try addressing #91: bail if blockstate's block isn't a crate.
This also changes a bunch of typed parameters (ArrayList) to interfaces
(List) for better accessibility of functions.

Further, it changes a number of functions to bail early if the list of
positions is empty.

Finally, it removes a few streams and replaces them with standard
if-loops as I've recently become aware of how significantly annoying
such things are if called frequently.

I'm not sure how frequently these were called, but it's still
potentially problematic.